WIP creating app settings screen
