+++ app/containers/addCabal.js
@@ -59,7 +59,7 @@ class addCabalScreen extends Component {
           placeholder='Pick a nickname' />
         <h2>
           Don't have a swarm to join yet?<br /><br />
-          <a href='#' onClick={this.newCabalPress.bind(this)}>
+          <a className='button' href='#' onClick={this.newCabalPress.bind(this)}>
             Create a Cabal
           </a>
         </h2>
