+++ app/styles/style.scss
@@ -63,6 +63,24 @@ form {
   padding: 0;
 }
 
+.button {
+  display: inline-block;
+  font-size: 0.875rem;
+  font-weight: 600;
+  border: 1px solid rgba(0, 0, 0, 0.15);
+  padding: 0.4rem 0.875rem;
+  border-radius: 3px;
+  color: rgba(0, 0, 0, 0.66);
+  background-color: #fff;
+  transition: all 0.075s ease-in-out;
+  cursor: pointer;
+  text-decoration: none;
+}
+
+.button:hover {
+  box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.1);
+}
+
 /* Layout: Container */
 .client {
   display: flex;
@@ -409,23 +427,6 @@ form {
     text-align: right;
   }
 
-  .button {
-    display: inline-block;
-    font-size: 0.875rem;
-    font-weight: 600;
-    border: 1px solid rgba(0, 0, 0, 0.15);
-    padding: 0.4rem 0.875rem;
-    border-radius: 3px;
-    color: rgba(0, 0, 0, 0.66);
-    background-color: #fff;
-    transition: all 0.075s ease-in-out;
-    cursor: pointer;
-  }
-
-  .button:hover {
-    box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.1);
-  }
-
   .channel-meta__other .channel-meta__other__share {
     margin-left: 1rem;
   }
@@ -695,28 +696,36 @@ form {
 }
 
 .modalScreen {
-  display: flex;
   background: #fff;
 
-  .modalScreen__close {
-    font-size: 2rem;
-    color: rgba(0, 0, 0, 0.5);
-    padding: 1.5rem;
-    position: absolute;
-    top: 0;
-    right: 0;
-    border: 0;
-    cursor: pointer;
-    transition: all 0.1s ease-in-out;
+  .modalScreen__header {
+    display: flex;
 
-    &:hover {
-      color: rgba(0, 0, 0, 1);
+    h1 {
+      margin: 2rem;
+      font-size: 2rem;
+    }
+
+    .modalScreen__close {
+      background-color: #fff;
+      font-size: 2rem;
+      color: rgba(0, 0, 0, 0.5);
+      padding: 1.5rem;
+      position: absolute;
+      top: 0;
+      right: 0;
+      border: 0;
+      cursor: pointer;
+      transition: all 0.1s ease-in-out;
+  
+      &:hover {
+        color: rgba(0, 0, 0, 1);
+      }
     }
   }
 
-  h1 {
+  .modalScreen__body {
     margin: 2rem;
-    font-size: 2rem;
   }
 }
 /*
